[IMP] knowledge: improve rename on click when article is untitled
Purpose:
--------

Previously, when a user clicked on the name of an article, the article
was renamed using the article's title in the body only if its name was
"untitled" in the user's language.

Now, an exception has been added: the article is also renamed if the name
of the article is "Untitled" (in English).

Task-3004122

Part-of: odoo/enterprise#32315