[IMP] planning: remove override of `_get_view` using alternative
Overriding `_get_view` should be avoided when possible,
when there is an easier alternative.

In this case, the method override can be avoided by
simply using the key `form_view_ref` to correctly
choose the form to use in that action.

In this case, since the result of `_get_view` is cached,
since odoo/odoo#99417,
it even leaded to a bug when calling `get_view` twice,
once with `{'force_email': False}`
and once with `{force_email: True}`,
as the context key `force_email` is not taken into account,
and `_get_view` would therefore return the same view for
the two possible values.

closes odoo/enterprise#32043

X-original-commit: 3bf4f17d414ab5cbff5c4600975dd07fa0c0d265
Signed-off-by: Denis Ledoux (dle) <dle@odoo.com>