[IMP] sign,sign_itsme: improve error messages
In sign_itsme we currently just receive the error that itsme returns us and we
show it in a dialog. However, these errors are not translated to the user's
language. This commit allows us to receive the error code from the IAP server
and map it to a more meaningful and translated error for the user.

task-2969605

closes odoo/enterprise#32033

Signed-off-by: Arnaud Joset <arj@odoo.com>