[FIX] account_automatic_transfer: fix a test
test_action_perform_auto_transfer is calling
action_perform_auto_transfer twice at the last day of the month.
This is not an issue as the transfer from the next month will just be
there empty until the next day.

The test was failing on the last day of the month as it expects a single
 call to action_perform_auto_transfer.

closes odoo/enterprise#32100

X-original-commit: 5cc69837345f2b52946cbe0a4adc23f58ffd7d73
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Nicolas Viseur <vin@odoo.com>