[FIX] mass_mailing: fix missing dependency for mailing domain
`mass_mailing_sale_subscription` does not have an explicit dependency on
`mass_mailing_sale`, but the `super()` call in `_mailing_get_default_domain()`
depends on it, otherwise it fails because there is no such method in `base` model.

No dependency means `mass_mailing_sale` is not guaranteed to be
installed, though!

<details>
   <summary>stacktrace</summary>

```
  File "/home/odoo/src/odoo/addons/mass_mailing/models/mailing.py", line 1199, in _get_default_mailing_domain
    mailing_domain = self.env[self.mailing_model_name]._mailing_get_default_domain(self)
  File "/home/odoo/src/enterprise/mass_mailing_sale_subscription/models/sale_order.py", line 12, in _mailing_get_default_domain
    domain = super()._mailing_get_default_domain(mailing)
AttributeError: 'super' object has no attribute '_mailing_get_default_do
```
</details>

In order to not change dependencies in a stable branch, the parent
domain from `mass_mailing_sale` is inlined, as there's not much
to duplicate.

The dependency can be changed in master later.

closes odoo/enterprise#32106

X-original-commit: 3628d820a6a37d21bde16db120dc5a9247f72b66
Signed-off-by: Olivier Dony <odo@odoo.com>
Signed-off-by: Arnaud Joset <arj@odoo.com>