[FIX] account_intrastat: fix report warnings/caret, hide transaction id
This commit is a combination of three fixes,
- Hide transaction id on the account move lines in the account move form view

The intrastat_transaction_id represents the type of transaction in the
intrastat system. This is provided at the level of the account move
line.

It has been decided that it should be optionally hidden by default. This
is as simple as changing the 'optional' attribute on the field.

- Fix name on intrastat report actions

Correct the name on the existing actions associated with the warnings
displayed on the intrastat report for misconfigured lines.

- fix caret options for intrastat report

The intrastat report lines have a caret that should provide the option
of visiting the invoice associated with the line of the report.

This commit specifies the _caret_option_initializer function on the
intrastat report custom handler, which overides the parent definition
and allows the 'View Journal Entry' caret option to be displayed from
the report line.

These fixes relate to bugprovements for v16:
https://pad.odoo.com/p/accounta_matata_v16

Part-of: odoo/enterprise#32379