[FIX] website_helpdesk_knowledge: delete helpdesk article
Prior to this fix, it was possible to move to trash an article
that was set on a helpdesk team as website_article_id.
As a consequence, clicking on 'Browse Articles', on the website help
page led to an empty page or an error.
Now, a validation error is displayed when trying to move to trash.

task-3002241

closes odoo/enterprise#32093

Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>