[FIX] hr_appraisal: goal field access fix
This commit fixes the way goal fields work in hr_appraisal,
as well as the role of appraisal officer:
- appraisal officers are now automatically set as employee officers
- someone without appraisal access can create goals for himself,
their subordinates as well as people having them as manager on an appraisal.
- only appraisal officers can modify the goal's manager
- a goal's manager is the employee's manager, but if no manager is set for the
employee, the user will be set as manager.
This commit also fixes the appearance of a traceback appearing while
searching through the employee list of either the employee or manager field.

task-2967514

closes odoo/enterprise#32094

Signed-off-by: Kevin Baptiste <kba@odoo.com>