[FIX] account_reports: include multivat variants properly in the tax report selector
Before this fix, in multivat setup, the foreign reports were not shown properly in the tax report selector, even we doing the right fiscal position setup.

This was due to a mismanagement of the fiscal_position option in the function computing the availability of variants. We fix that by not depending on the value set for this fiscal_position option in that function anymore.

Signed-off-by: Julien Alardot (jual) <jual@odoo.com>
X-original-commit: 0818be916151fdf92a5fd1a495f03937a94fb490
Part-of: odoo/enterprise#32311