[FIX] data_merge_model: see notification message from find_duplicates
With this commit, we are able to see the message generated by the cron "Data Merge:
Find Duplicate Records", even if the user preference for notification is set to
Handle in Odoo.

Previous behavior:
- Your user -> Preferences -> Notifications -> Handle in Odoo
- Contact, create two duplicates (I created two contacts named test with the email
test@test.fun)
- Data cleaning -> Configuration -> Deduplication -> Contact -> Notify Users, set to
your user and to 1 day
- Scheduled Actions -> Data Merge: Find Duplicate Records -> Run Manually
- Your notification bar now has 1 more notification, which you can't access other
than by entering the Discuss app and seeing your inbox

This behavior was ver confusing for customers.

Behavior now:

You can see this notification as well as the message associated in the discuss bar.

OPW-2903716
OPW-2992342
OPW-2634116

closes odoo/enterprise#32170

X-original-commit: babefcbb99871332c8f8e34133ab5702066ad758
Signed-off-by: Kevin Baptiste <kba@odoo.com>