[FIX] {snailmail_}account_followup: fix followup line bugs and wizard layout
The followup rework didn't implement the case when a company would have no followup line defined.
This would lead to tracebacks when trying to manually followup partners for such companies.

When joining invoices to a followup letter, every unpaid invoice would be attached, including those that are supposed to be blocked.

Some CSS changes broke the display of the followup wizard, rendering it unreadable.

This commit fixes those issues by:
- fixing the followup process when no followup lines are defined (allowing a similar flow, but with default values instead of the followup line's values)
- only including invoices that are not blocked when joining them to a followup letter
- fixing the followup wizard layout (mainly reworking the colspan for groups)

Part of the big task fixing bugs in accounting
Task id=2996493

closes odoo/enterprise#32339

X-original-commit: 7086f39f16bf849bbbb632572faf8e185743a55a
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Popeler Antoine (anpp) <anpp@odoo.com>