[FIX] industry_fsm_stock: fix delivery update
This commit purpose is fix the delivery update. When more than one fsm
product is on an SO, and task were created for those sol, if products
are added to different tasks a common delivery is created for the so
with all the products. Currently, when one of the task is marked as done,
the delivery is not correctly update, nor are the sol with the material
products related to the task.

the commit:
 - correctly set the quantity done for the sol and the delivery when a
task is marked as done
 - Once all the tasks composing the sol are marked as done, the delivery
is also set to done.
 - add a test to ensure future modification keeps this behavior

task 3000244

closes odoo/enterprise#33272

X-original-commit: 70637e35511b87c7308f238268b7f35e272c0683
Signed-off-by: Xavier <xbo@odoo.com>
Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>