[FIX] sign: signature not being overwritten when drawing
In sign, we try to avoid as much as possible overwriting the signature of the
user, as the device which was used to sign affects the resolution of the image
saved in the user profile. Therefore, we only save the signature when the user
interacted with any of the buttons in the signature widget. However, the
signature was not being changed if the user didn't click on any of the buttons
but drew while being on the draw mode. This commit fixes this.

task-3002677

closes odoo/enterprise#33855

X-original-commit: d16c9de793ba2971367dcc9e8513e15f5a13974c
Signed-off-by: Arnaud Joset <arj@odoo.com>