[FIX] account_invoice_extract: revert to old tax selection logic
The adaptation of the code handling the invoice lines creation
following the accounting rework (see odoo/enterprise#29758) was
flawed.
It didn't take into consideration that some taxes could be set by
`account_predictive_bills`.
In that case, the OCR needs to check afterwards that the taxes
set are compatible with what was detected. In the old version,
the prediction was done after the invoice lines creation, so no
check could be performed.

After this commit, the tax logic should be equivalent to what was
done before the accounting rework.

closes odoo/enterprise#32199

X-original-commit: 1987c233cf1bc10a812ac69139476d98353a1f79
Signed-off-by: Florian Daloze (fda) <fda@odoo.com>
Signed-off-by: Louis Baudoux (lba) <lba@odoo.com>