[FIX] sale_renting: rental wizard opening on sales and subscription
The rental configurator wizard was triggered in 'Sales' and 'Subscription'
app when a rentable product was added to a sales order line.

This should only happen in the 'Rental' app.

This commit makes sure the wizard is only shown in the 'Rental' app, or
eventually in other apps when editing a rental line.

Task - 3002268

closes odoo/enterprise#32164

Signed-off-by: Victor Feyens (vfe) <vfe@odoo.com>