[FIX] hr_payroll: change full time and work time rate layout
Since the form is always in edit mode, the units of the full time and work time rate fields
on the resource calendar form are pushed on the right side.

This commit reduces the size of these fields to have the units closer to the value of the fields.

task-2996235

closes odoo/enterprise#32539

X-original-commit: 703b2b2fbb9b49c46495ba1c0786105e51b50c22
Related: odoo/odoo#102711
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>
Signed-off-by: phwa-odoo <phwa@odoo.com>