[FIX] knowledge: properly set the name of /kanban and /list
When the user uses the /kanban or the /list command and enters a name
in the dialog, the embedded view ignored the name entered by the user
and used the name of the corresponding action. This commit will ensure
that embedded view will get the name entered by the user.

task-3004117

X-original-commit: 5eb5da41c6b4e9dd3fd8df03ffac4b7df4465510
Part-of: odoo/enterprise#32543