[REF] *: GanttView: stop using legacy FormViewDialog
This commit removes the last usecases of the legacy FormViewDialog,
which is about to be removed. It replaces them by the wowl
FormViewDialog.

Some legacy tweaks/hacks needed to be implemented in a different
way as they can no longer work in the new infrastucture. For
instance, listening to the custom event `execute_action` triggered
on the dialog is no longer possible. For that reason, we implemented
the logic in the form js_class instead (the one inside the dialog).

X-original-commit: 41e1316c8cd4f2e82ff3b06895b0d7405b3e3efd
Part-of: odoo/enterprise#32281