[FIX] knowledge: fix refresh issues after move to
When moving an article from one location to another, many fields may need to be
re-évaluated.
This commit force the record to be updated at client side so the display can
be adapted in consequence (attrs, t-if, etc..).

Also, This commit force to save before calling the moveTo in order to keep the
changes that have been done before calling the moveTo that were not saved yet.
Tipically, if you type something in the html field and directly move the article
somewhere else.

Task-3004047

X-original-commit: 56f1636d02884670df36c763af8f6ee4d170e4c8
Part-of: odoo/enterprise#32439