[PERF] knowledge: use state to control chatter display and lessen RPCs
Currently chatter is always displayed and made invisible while the panel is
not activated. This however has the side effect of triggering a lot of
unnecessary calls when switching articles :

  * one RPC due to html field committing itself (update old chatter);
  * one RPC due to save notifying models that reload (update old chatter);
  * one RPC due to loading the chatter of the new article (fetch new chatter);

By correctly using the state to control chatter display in most cases all
those RPC are removed. Until the panel is not activate, the chatter itself
is not present in the DOM and no extra RPC is made.

Task-3004723

Part-of: odoo/enterprise#32192