[FIX] knowledge: autofocus inside the body
Purpose:
--------
When opening an article, the focus is not set inside the body of the
article anymore.

Now, when opening an article, the focus is set inside the body of the
article, only if the body of the article is empty: `focus()` places the
cursor at the beginning of the body, which is not handy when the article
already has some content.

Task-3004122

Part-of: odoo/enterprise#32263