[IMP] sale_subscription: allow to change customer of confirmed orders
Before this commit, it was not possible to easily change the customer of a
confirmed subscription. This commit introduces a wizard to do it.
We chose to create a dedicated wizard and not the form view because
it is a flow that should not happens a lot and it allows to easily
control who can do it. The regular partner_id field remains editable
under the existing conditions: draft or sent quotation.

Task-id: 2988345
Part-of: odoo/enterprise#32591