[FIX] documents_{project,fsm}: fix automatic folder creation
This commit fixes two problems in the automatic generation and deletion of
project folders:

- In `documents_project`, the folders are generated based on the domain defined
in the `_get_basic_project_domain` method. This can create inconsistenties when
installing documents while fsm is already installed, where a fsm project will have
`use_documents` to True, but no folder will be generated.
- In `documents_fsm`, the folders are removed for fsm projects but the `use_documents`
field is not set to False, which causes an error when editing those projects because
the folder is then required but not set

closes odoo/enterprise#32216

X-original-commit: f6c3f9a97f3f6887535546a9f3126f1aa4c2fac1
Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>
Signed-off-by: Larcin Vincent (vila) <vila@odoo.com>