[FIX] timesheet_grid: expand grid to last sunday
Steps :
Say today is Friday the 7th.
Go to my timesheets, and encode a timesheet on a new line to monday the 3rd.
Go to next week.
We expect a line to be displayed for this project/task on next week.

Issue :
This line is not displayed.

Cause :
We look for timesheets in the past 7 days, whereas we should look for
timesheets from sunday of last week to today.

task-3002629

closes odoo/enterprise#33357

X-original-commit: ae6c2dcd215ba140b3cf94e431359bd353a07d6f
Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>
Signed-off-by: Xavier <xbo@odoo.com>