[FIX] timesheet_grid: expand grid to last sunday
Steps :
Say today is Friday the 7th.
Go to my timesheets, and encode a timesheet on a new line to monday the 3rd.
Go to next week.
We expect a line to be displayed for this project/task on next week.

Issue :
This line is not displayed.

Cause :
We look for timesheets in the past 7 days, whereas we should look for
timesheets from sunday of last week to today.

task-3002629

closes odoo/enterprise#33331

X-original-commit: 81479ec7b1750c3b7811b35dbed5f47056aa48a6
Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>