[PERF] knowledge: display share panel on demand and lessen RPCs
Currently each loading of an article adds an RPC linked to permission panel.
It calls``get_article_permission_panel`` to have permissions and members data
to display in the panel.

However this panel is toggled on demand and not displayed by default. We can
therefore delegate the RPC at the click, allowing to lessen number of RPCs
done when loading an article.

Publish flag management in 'website_knowledge' is also improved by this
commit and the chatter update. Previously to those commits, toggling the
button was generating 9 RPCs while we have now 2 RPCs.

Task-3004723

Part-of: odoo/enterprise#32192
Co-authored-by: Aurélien Warnong <awa@odoo.com>
Co-authored-by: David Beguin <dbe@odoo.com>