[IMP] timesheet_grid: improve the validation process of an approver
Before this commit, the "My Team" filter is showing all the users
that are part of a project from wich the user is manager or timesheets
of employees for which he is the responsible.
Also, a user who has the "User: All Timesheet" access right set,
should be able to validate timesheet of other users that do not have
a timesheet approver and a parent set.

This commit updates the filter condition to only see the timesheets
linked an employee for which the current user is his timesheet manager
or his manager. It also changes the validation process for the
Timesheets approver, he can now validate the timesheets linked to an
employee for which the current is the manager or the timesheet approver
or that employee has no manager and no timesheet approver.

task-3061362

closes odoo/enterprise#34181

Related: odoo/odoo#108060
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Co-authored-by: Xavier Bol <xbo@odoo.com>