[IMP] sign_itsme: use correct iap api endpoint
After deployment to production, several aspects of the iap service were
modified to better integrate within the iap ecosystem, including its
endpoint.

This commit makes sure that users need not define a system parameter for
the integration to work - the parameter is used e.g. to switch to
sandbox, etc.

closes odoo/enterprise#32248

X-original-commit: 997764bdb86e5d03b67dbe4cee7a450ac664b173
Signed-off-by: Bouvy Damien (dbo) <dbo@odoo.com>