[IMP] sign_itsme: use correct iap api endpoint
After deployment to production, several aspects of the iap service were
modified to better integrate within the iap ecosystem, including its
endpoint.

This commit makes sure that users need not define a system parameter for
the integration to work - the parameter is used e.g. to switch to
sandbox, etc.

closes odoo/enterprise#32226

Signed-off-by: Bouvy Damien (dbo) <dbo@odoo.com>