[FIX] sale_subscription: fix product's descriptions not being correctly copied to SOL
Currently in sale_subscription, when using a quotation template composed of
multiple lines with the same product but different descriptions, the
descriptions are not properly copied. They are all set to the first product's
description in the quotation template. This commit fixes this by removing the
override of _compute_name which adds some unrequired dependencies causing the
method to be called when it's not supposed to.

task-3003962

closes odoo/enterprise#32237

Signed-off-by: Arnaud Joset <arj@odoo.com>