[FIX] account_conso: Fix the copy of a consolidation
Consolidations were being copied as such: the accounts were copied but
not the groups. As such, the new accounts were linked to the groups
from the copied consolidation which is completely wrong.

Now, we will change the way it is copied as such:
=> Duplicating the account is copying the groups
=> Duplicating the group is copying the accounts
So now, we have the correct chart, with the correct groups and accounts
=> Finally, link the account with the chart as it's not done yet

opw-2985652

closes odoo/enterprise#32748

X-original-commit: ea2383e34caf491ff5d9b57344265cd5269127f5
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Nicolas Viseur <vin@odoo.com>