[PERF] knowledge: update article members once when inviting people
Previously to this fix update of members was done in two separate steps when
invite members, once for external people, once for internal people.

With this fix, we concatenate commands to update members so that write is
called once on the article. We simply extract the command generation from
``_add_members`` method into its own sub method that can be called directly.

Task-3004723

X-original-commit: bdff19f768a7b282d7b9afc499ee3d68f5b0eb16
Part-of: odoo/enterprise#32450