[PERF] knowledge: update article members once when inviting people
Previously to this fix update of members was done in two separate steps when
invite members, once for external people, once for internal people.

With this fix, we concatenate commands to update members so that write is
called once on the article. We simply extract the command generation from
``_add_members`` method into its own sub method that can be called directly.

Task-3004723

Part-of: odoo/enterprise#32192