[FIX] delivery_easypost: multi pack label issue
Steps to reproduce:
- create a shipping method with easypost
- have credentials with interlink express
- create a sale order with two products and confirm it
- in delivery, put the quantity done for the first product - Put in pack | put the qty done for the second product - put in pack
- validate

Issue:
None type is not subscriptable

Cause:
With some carriers, such as Interlink Express, they put the labels information in the first shipment object.

Solution:
Iterate through `postage_label` only if it exists. It won't impact the flow. The only differnence is that for Interlink Express there will only be one pdf (with multiple labels) in the chatter instead instead of multiple pdf's (with one label) such as for other carriers.

opw-2890269

closes odoo/enterprise#33265

X-original-commit: 1e40b9a01ac7781dc61e95e260a71936cef99f52
Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>
Signed-off-by: Yolann Sabaux (yosa) <yosa@odoo.com>