[IMP] knowledge: use new emoji list for random emoji picker
Purpose:
--------
Instead of using the old and small list of emojis for picking an emoji
at random when clicking on the "Add icon" button, use the new one which
contains much more emojis, and which is currently used for the emoji
picker.

The list of disallowed emojis has been updated accordingly, removing some
emojis of it as the chances of picking them at random are very small, but
adding some that we really do not want to show (eg. the middle finger).

Task-3004122

Part-of: odoo/enterprise#32315