[IMP] knowledge: add tests for 'get_valid_parent_options' and 'copy_batch'
In this commit we add unit and performance tests for get_valid_parent_options
that is used in moving article dialog. As this may be used quite often let us
be sure it works as intended and add performance tests.

We also improve tests for ``copy_batch`` (check descendants, add more tests
as admin that has more rights), and add performance tests for that method
to try to optimize it a bit afterwards.

Task-3004723

X-original-commit: 93a13911a3077d517c2372f00c2afbefaf2c8235
Part-of: odoo/enterprise#32450