[FIX] account_reports: Fix default variant report selection
When a report is loaded, the variant selected was the first found,
not the one corresponding to the company country

In most cases there is only one variant allowed so it loads
the correct one. But when several variant are allowed we want
to be sure that the first to be loaded is the company's country one
if any. If none, we can then load the first found as there is no way
to know which should be the first

closes odoo/enterprise#32311

X-original-commit: 784c08070253021099285ac8d377721bdfe20fc8
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Julien Alardot (jual) <jual@odoo.com>