[FIX] sale_subscription: prevent crash when compute qty_invoiced
Before this commit, we could have historical data of
account move line of subscription without subscription_start_date
or subscription_end_date set. As these values are copared to date for the
subscription, a crash would occurs. This commit make sure to compare
non-falsy values or fallback on the move invoice_date value to get the
invoiced qty.

X-original-commit: cbbd8aab75e44eea46f9d3f7f24c091e06989b96
Part-of: odoo/enterprise#32572