[PERF] knowledge: improve performances of copy_batch
Knowledge article has en option available for administrators allowing to copy
articles in batch from the list view. This helps admins creating default
hierarchies, used as templates.

In this commit we make ``copy_batch`` copy implementation of ``Model.copy()``
instead of calling it directly. Indeed base copy() is still a per-record
implementation, and we could create records in batch. Other sub-calls (like
``copy_data()`` and ``copy_translations``) are still calling the singleton
version of ORM to avoid duplicating the ORM. Optimizing the ORM is outside
the scope of this PR.

We also limit searches when duplicating articles. Because filter_domain with
child_of didn't seem to work, we instead use the code to extract ancestor_ids
by making it a model helper method. Tests are added to strengthen code.

Task-3004723

Part-of: odoo/enterprise#32192
Co-authored-by: Florian Charlier <flch@odoo.com>