[FIX] account_reports: fix missing caret options on partner ledger
The caret options are currently missing from the partner ledger. These
normally allow the user to visit the appropriate account_move or payment
associated with a specific report line.

In this commit the _caret_options_initializer function is defined on the
partner ledger custom handler. This overrides the definition in
AccountReportCustomHandler (which would return an empty dictionary), and
returns a dictionary containing a mapping from the model on which the
caret option is to be displayed to the name of the caret option and the
specific action it performs. This allows for the caret options to be
displayed by the report handler.

closes odoo/enterprise#32270

Signed-off-by: Olivier Colson (oco) <oco@odoo.com>