[FIX] spreadsheet_edition: use custom ModelFIeldSelector
We require a patched version of `ModelFieldSelector` in order to expose
the fieldName chain as well as the fields it contains in order to
validate its consistency with the related model of a filter.

This commit exposes the said fields and introduces the possibility to
target multidepth fields (only in debug mode) for finer filtering.

task 2992752

closes odoo/enterprise#32526

X-original-commit: 46c8f08ba6a22c865a0475939843e15343f68c8a
Related: odoo/odoo#102688
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>
Signed-off-by: Pierre Rousseau (pro) <pro@odoo.com>