[FIX] spreadsheet_edition: use custom ModelFIeldSelector
We require a patched version of `ModelFieldSelector` in order to expose
the fieldName chain as well as the fields it contains in order to
validate its consistency with the related model of a filter.

This commit exposes the said fields and introduces the possibility to
target multidepth fields (only in debug mode) for finer filtering.

task 2992752

closes odoo/enterprise#31633

Related: odoo/odoo#100825
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>