[FIX] knowledge: add module png icon
Purpose:
--------
Recently knowledge's icon has been changed to a svg icon instead of a png.
However, the app module works with png icons, and therefore knwoledge
didn't have an icon in the app module.

The PNG icon has been re added so that the icon is shown in the apps
module.

Task-3004122

X-original-commit: 219f1a63416090c217a356150eb3790ff4592c86
Part-of: odoo/enterprise#32334