[FIX] sale_subscription: update next_invoice_date in manual invoicing
Current behaviour:

When we have a payment token linked to the SO, we skip the update
of the `next_invoice_date` to update it on the reconciliation of
the payment.
The issue is that the reconciliation call back is only called when
the invoice transaction is created by `do_payment` which only happen
for automatic invoice.

Behaviour after PR:

If we are creating a new manual invoice, the next invoice date is
updated even if we have a payment token

task-id : 3007712

X-original-commit: 9ef3dfac1338062f43726b2ee1d36cbe408f7f6f
Part-of: odoo/enterprise#33995