[PERF] knowledge: handle property panel display using owl
Use OWL management to display or hide the property panel instead of doing
low-level DOM manipulations. Implemented flow is the following

  * instantiate a kanban view for articles using the /kanban view;
  * create an item;
  * go on top of article to display actions, click on "Properties";

This will at the same time display the properties wheel icon and toggle it
as it was clicked by the user. It can then be toggled on demand.

Task-3004723

Part-of: odoo/enterprise#32192