[IMP] {website_}knowledge: handle website conflicts
Purpose
=======
Improve the UI and the redirection for logged in users.

Specifications
==============
Hide the livechat button from knowledge since it
should not appear nor be accessible.

Redirect anonymous users on the login page if
they are trying to access an existing but
unpublished article.

To prevent a broken UI look, adapt the size
of the portal users virality message and hide it on
mobile devices.

Task-2936605

closes odoo/enterprise#32556

X-original-commit: 1b4e3898f7225fc620563fac06b47886f533da68
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>