[FIX] timesheet_grid: remove timer funciton from time-off entries
This commit removes the timesheet timer from the timesheet
entries that are related to time-off. This removes the
ability of the user to edit the timesheeted amount by running
the timer.

This is a small design overlook that now should make the
timesheet app more robust to unexpected user input.

task-3007842

closes odoo/enterprise#32545

X-original-commit: 85fbbc645d598943323c1a3ec3c9463025bcc189
Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>