[IMP] knowledge: prevent closing share panel on invite modal click
Purpose:
--------
Previously, when inviting someone to an article, a click anywhere on the
screen (including on the invite modal and its fields) closed the share panel.

Now, the share panel cannot be closed when the invite modal is shown, so that
the user can see the members that already have access to the article,  and
know which other people he should send an invitation to.

Task-3004122

Part-of: odoo/enterprise#32263