[FIX] account_reports: round other engines' totals when computing aggregations
When computing the report, engines don't round the amounts they return. Because of that, some amounts that should be 0 will instead be very small floats (like 1e-20, for example). If those amounts needed to be used in an aggregation, this failed when trying to expand the aggregation formula, because the "-" before the "e" was identified as an operator, and the terms were split in such a way that 1e could not be evaluated.

We solve that by rounding all the amounts obtained from other expressions, as they will be rounded in the report anyway. We also avoid futher rounding issues doing so.

closes odoo/enterprise#32364

X-original-commit: df6d1236d9d4c0abb71ead7e7034ce8daee05fe8
Signed-off-by: Laurent Smet <las@odoo.com>
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>