[FIX] project_timesheet_forecast: fixed trackback while accessing task_id
In planning.slots, when generating timesheets through the list view,
There is no task_id in the timesheet_read_group since task_id is a removed in
task-2941828. As a result, it generates a traceback when generating timesheets.

By this commit, task_id has been removed from the group in timesheet_read_group
and during the calculation of timesheet_date.

task-2997424

closes odoo/enterprise#32337

Related: odoo/odoo#102305
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>