[FIX] documents: fix impossible to upload files via upload button
During the documents refactor to wowl2 the upload function was
wrongfully made to get `tag_ids` as an optional argument.
However the route used to upload files does in fact require the
argument to work.
During the cleanup the `tag_ids` argument was moved from `params` to the
main object used call a file upload, since `params` was designed to be
used for optional arguments (which there are none now and has been
removed), however one of the call to the uploadFiles function was not
properly adapted and still used the old optional params to provide the
tags, this results in the upload button being completely unusable.

TaskId-3008459

closes odoo/enterprise#32373

X-original-commit: f51597a302589ad1038f4d16b2cabc597dcfee75
Signed-off-by: William Braeckman (wbr) <wbr@odoo.com>