[IMP] knowledge: remove custom editor style and tiny design imps
Purpose:
--------
Remove scss rules for knowledge's editor: it should offer the same design
as any pad.

Replaced the space character inside the "create" and "share" buttons with
a small margin-end, to prevent misalignments occurring with translations
that removed this space.

Added an emoji in the list of disallowed emojis for the random emoji
picker.

Task-3004122

Part-of: odoo/enterprise#32263