[FIX] knowledge: don't show favorited article items in normal article tree
Before this commit, once an article item was favorited, it was also displayed
in the normal article tree (e.g.: in workspace section).
This commit hide the article items from the standard article tree, even if they
are favorited. Favorited article items are still displayed in favorite tree.
Also, favorited article items are directly added to the favorite tree right
after beeing favorited, and not only after reload like it was the case before.

Task-2999302

closes odoo/enterprise#32439

X-original-commit: c084716286322dd42fd922891127608615b19f03
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>