[FIX] sale_subscription: Only reconcile payment on auto_invoice
The payment should only be reconciled in auto_invoice mode.
If we are not in auto invoice mode, the payment will be saved and
will be reconciled when the invoice is created by the cron.

Before this commit, the payment would be reconciled automatically
and this would increase the next_invoice_date without actually
creating an invoice (even if the transaction was existant).

closes odoo/enterprise#32410

X-original-commit: b23ee529c6c5a0f76037a10d9605621ca332fcad
Signed-off-by: Arnaud Joset <arj@odoo.com>