[FIX] stock_barcode: bus off listener
In the MainComponent `onMounted`, we add set some event listeners on the
bus but we don't remove them once when the component is unmounted, which
can cause some issue. For example:
- Go to Barcode > Operations > Receipts and create a new receipt;
- Go back and create another new receipt;
- Scan a product -> a new line is created.
- Click on this new line's edit button -> traceback.

As the event listeners weren't shut down, when the LineComponent
triggers an 'edit-line' event, both the previous and the current
MainComponent will catch it and run the corresponding function, so the
first time, it will search for the edited line in the previous env's
model, which has not it, which causes the traceback afterward.

task-2960938

X-original-commit: 6cd1aeb8365c27473f0d0e13d56cdbab2e7654cd
Part-of: odoo/enterprise#32481