[FIX] appointment: fix input display
Purpose
=======
Fix the inline input fields which
width seems broken because of the new
always edit mode.

Specifications
==============
In the appointment form view, limit the width
for the float inputs and the integer input to
fix the inline display and prevent from having
huge white gaps.

Task-2996467

closes odoo/enterprise#34402

X-original-commit: 8334aa977033aea19801aba93941153613cd0b41
Related: odoo/odoo#106570
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>