[FIX] stock_barcode{,_picking_batch}: scan package for several SML
When scanning a package, if there are several SML with the same product,
only one of these SML will be updated.

To reproduce the issue:
1. In Settings, enable "Packages"
2. Create a storable product P
3. Update its quantity: 10 x P in a package PK
4. Create, confirm and assign two planned delivery orders D01, D02
    - With 3 x P
    - With 7 x P
5. Create and confirm a batch transfer BT with D01 and D02
6. Barcode > Batch Transfers > BT, scan PK

Error: the quantity of one line is equal to 10 and the quantity of the
other one didn't change

In the Barcode app, when processing a package, we iterate on each quant,
search for a related line and update it. In the above case, the package
only has one quant (10 x P). It finds one of the lines and then simply
updates it. When we find a line, we should fulfill its need (nothing
more) and retry the search & update step with the remaining quantity.

OPW-2976150

closes odoo/enterprise#33593

X-original-commit: 5ac2d0ae2ce35530d54ba6800be519a08051242e
Signed-off-by: Steve Van Essche <svs@odoo.com>
Signed-off-by: Adrien Widart <awt@odoo.com>