[FIX] account_reports: partner ledger: make proper use of _query_get()'s return
A query used by the partner ledger report was not making use of the tables returned by the query_get(), but was using its where_clause. Because of that, depending on the content of the options dict, it was possible that the where_clause referenced joined tables that did not exist (for example, account_move_line__move_id, if any domain computed from the options needed to access a field on the lines' parent move), leading to a crash when trying to run the query.

closes odoo/enterprise#32406

Signed-off-by: Laurent Smet <las@odoo.com>