[FIX] data_*: adapt form views for grid
Before this commit, the frequency fields were not aligned. This is due to the
new CSS grid system in the form views for inner groups.

This commit reorganize the arch itself to have the the relevant fields and their business
modifiers in the same cell of the inner group. It not only makes sense technically, it does so
in business terms as well.

*: cleaning, merge

closes odoo/enterprise#32475

X-original-commit: 9091095172cb82be5cf914964de20fc1c4c3b845
Related: odoo/odoo#102578
Signed-off-by: Adrien Dieudonné (adr) <adr@odoo.com>