[FIX] data_*: adapt form views for grid
Before this commit, the frequency fields were not aligned. This is due to the
new CSS grid system in the form views for inner groups.

This commit reorganize the arch itself to have the the relevant fields and their business
modifiers in the same cell of the inner group. It not only makes sense technically, it does so
in business terms as well.

*: cleaning, merge

closes odoo/enterprise#32407

Related: odoo/odoo#102435
Signed-off-by: Adrien Dieudonné (adr) <adr@odoo.com>