[FIX] social: restore campaign form access.
When trying to access a campaign with posts (or push notifications),
a traceback appears. This is because the posts are in kanban mode in
the embed, but when the template is rendered, the context does not know
of the function formatPost used in the template.

Therefore, display the record message without formating it if that
function is undefined. The same is done for when clicking the image
carousel: nothing will be done instead.

Task-3016141

Part-of: odoo/enterprise#32715