[REF] project_enterprise: review assets for project sharing
Before this commit, the assets used in project sharing takes almost all
assets in `web.assets_backend`, this way is not a good idea since only
the kanban view and list are used and the chatter is not the backend
one. So the assets could be filtered to minize the size of the bundle
for project sharing to also reduce the loading when the user goes to
project sharing feature.

This commit reviews the assets used in the project sharing to reduce
the size.

task-2947516

Part-of: odoo/enterprise#32486