[REM] *: remove legacy views converted to owl
This commit removes legacy views that are not used anymore as they
have been converted to owl. They could actually have been removed
in the commit that converted them, as we never needed to have a
double implementation of views, unlike we need to (field) widgets.

In the community PR odoo/odoo#102540 removes the form and list views
from the view registry, so this commit directly imports them in
Studio. Those two views are a special case in v16, since we use the
owl implementation outside studio, but still the legacy one in
studio.

closes odoo/enterprise#32501

X-original-commit: 4c85cc09a72ce4c09547f6dde347ab804273b19c
Related: odoo/odoo#102634
Signed-off-by: Géry Debongnie <ged@odoo.com>
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>